+++ c/smalloc.h
@@ -1,8 +1,8 @@
 #ifndef __SMALLOC_H__
 #define __SMALLOC_H__
 
-void* smalloc(unsigned long size);
+void* smalloc(unsigned int size);
 void sfree (void *ptr);
-long GetMallocNum ();
+int GetMallocNum ();
 
 #endif
